# models/community.go
package models

import "time"

type Community struct {
	ID   int64  `json:"id" gorm:"column:community_id"`
	Name string `json:"name" gorm:"column:community_name"`
}

type CommunityDetail struct {
	ID           int64     `json:"id" gorm:"column:community_id"`
	Name         string    `json:"name" gorm:"column:community_name"`
	Introduction string    `json:"introduction,omitempty" gorm:"column:introduction"` // omitempty没有就不展示
	CreateTime   time.Time `json:"create_time,omitempty" gorm:"column:create_time"`
}

func CommunityTableName() string {
	return "communities"
}